backend/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.5.7
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Django==3.2.6
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
heroku==0.1.4
idna==3.2
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.1
pycodestyle==2.7.0
PyJWT==2.1.0
python-dateutil==1.5
pytz==2021.1
regex==2021.8.3
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
urllib3==1.26.6
psycopg2-binary
